specify borough when querying coords; was getting matches outside of manhattan
I forgot that the other boroughs have some of the same street names and addresses as manhattan. i updated the code to pull the 'county' (ViolationCounty) from the data frame and pass it into the query

according to api website, i may need to use actual borough names, and not the county. Though the website says that county name of "NY" or "New York" is used to mean "Manhattan", so I should be okay for this particular run. For others, (1=Manhattan, 2=Bronx, 3=Brooklyn, 4=Queens, 5=Staten Island)