fixed bad merge due to DoW mismatch
i had merged on the wrong DoW convention -- python has monday as 0, sunday as 6; in R, I had sunday as 1 and saturday as 7 (i think?). This screwed up the merge, causing duplicates whenever a ticket was issued (whenver there was an entry in the original data set).

a side effect of this bad merge was that every ticket was preceded by a 0 value ... which is probably why glm() had been throwing those warnings

thus, i updated model AIC and correlations between fitted and observed

assessment of path forward remains the same